requirements.txt
flask
flask-cors
flask-sqlalchemy
flask-jwt-extended
google-generativeai
python-dotenv
werkzeug
